backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
firebase-admin==6.5.0
google-cloud-documentai==2.26.0
google-generativeai>=0.5.0
email-validator>=2.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.7 # Added for form data/file uploads